+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Allow user to set Acq method EDIFACT export triggers for POL. Refs UIOR-371.
 * Translations for orders lines filter by export date. Refs UIOR-883.
 * Do not export order lines where Manual is true for the PO. Refs UIOR-890.
+* Disable Check for potential duplicate order lines in Settings. Refs UIOR-888.
 
 ## [3.0.3](https://github.com/folio-org/ui-orders/tree/v3.0.3) (2021-12-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.2...v3.0.3)